refactor file arrangement
